Started upgrading to Polymer 3
