+++ .travis.yml
@@ -1,9 +1,7 @@
 language: node_js
 node_js: stable
 sudo: required
-before_script:
-  - npm install -g polymer-cli
-  - polymer install --variants
+before_script: npm install -g polymer-cli @advanced-rest-client/wct-istanbub
 addons:
   firefox: latest
   apt:
@@ -13,16 +11,11 @@ addons:
       - google-chrome-stable
   sauce_connect: true
 script:
-  - xvfb-run polymer test --plugin local
+  - xvfb-run polymer test --module-resolution=node --npm --plugin local
   - >-
-    if [ "${TRAVIS_PULL_REQUEST}" = "false" ]; then polymer test --plugin sauce
-    --job-name "cookie-parser:${TRAVIS_BRANCH}"
-    --build-number=${TRAVIS_BUILD_NUMBER}; fi
-cache:
-  directories:
-    - node_modules
-after_success:
-  - node tasks/ci.js
+    if [ "${TRAVIS_PULL_REQUEST}" = "false" ]; then polymer test
+    --module-resolution=node --npm --plugin sauce --job-name
+    "cookie-parser:${TRAVIS_BRANCH}" --build-number=${TRAVIS_BUILD_NUMBER}; fi
 env:
   global:
     - secure: >-
