+++ package.json
@@ -1,8 +1,18 @@
 {
-  "name": "@advanced-rest-client/cookie-parser",
-  "version": "2.0.0",
-  "license": "Apache-2.0",
   "description": "A javascript cookie parser for javascript HTTP clients.",
+  "keywords": [
+    "javascript",
+    "cookies",
+    "HTTP"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/jarrodek/cookie-parser.git"
+  },
+  "homepage": "https://github.com/jarrodek/cookie-parser",
+  "name": "@advanced-rest-client/cookie-parser",
+  "version": "3.0.0-preview.1",
+  "license": "Apache-2.0",
   "scripts": {
     "test-sauce": "polymer test --plugin sauce --job-name \"cookie-parser:local-test\"",
     "test": "polymer test --plugin local",
@@ -12,9 +22,17 @@
   },
   "devDependencies": {
     "@polymer/gen-typescript-declarations": "^1.6.1",
-    "chai": "^4.1.2",
-    "mocha": "^4.1.0"
+    "chai": "^4.2.0",
+    "mocha": "^5.2.0",
+    "@webcomponents/webcomponentsjs": "^2.0.0",
+    "@polymer/test-fixture": "^4.0.2",
+    "sinon": "^7.2.3",
+    "wct-mocha": "^1.0.0"
   },
-  "private": false,
-  "main": "parser.js"
+  "main": "parser.js",
+  "contributors": [
+    "Pawel Psztyc",
+    "The Advanced REST client authors <arc@mulesoft.com>"
+  ],
+  "dependencies": {}
 }
