+++ polymer.json
@@ -1,7 +1,8 @@
 {
+  "npm": true,
   "lint": {
     "rules": [
-      "polymer-2"
+      "polymer-3"
     ]
   }
 }
