+++ test/merge-cookies-test.html
@@ -17,14 +17,16 @@ the License.
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-  <script src="../../webcomponentsjs/webcomponents-loader.js"></script>
-  <link rel="import" href="../../url/url-polyfill.html">
-  <script src="../../web-component-tester/browser.js"></script>
-  <link rel="import" href="../cookie-parser.html">
+  <script src="../../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+  <script type="module" src="../../../url/url-polyfill.js"></script>
+  <script src="../../../wct-browser-legacy/browser.js"></script>
+  <script type="module" src="../cookie-parser.js"></script>
 </head>
 
 <body>
-  <script>
+  <script type="module">
+  import '../../../url/url-polyfill.js';
+  import '../cookie-parser.js';
   /* global assert, Cookies */
   suite('Parse cookies with base URL', function() {
     this.timeout(2500);
