+++ test/parse-cookies-test.html
@@ -16,13 +16,15 @@ the License.
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1.0, user-scalable=yes">
-  <script src="../../webcomponentsjs/webcomponents-loader.js"></script>
-  <link rel="import" href="../../url/url-polyfill.html">
-  <script src="../../web-component-tester/browser.js"></script>
-  <link rel="import" href="../cookie-parser.html">
+  <script src="../../../@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+  <script type="module" src="../../../url/url-polyfill.js"></script>
+  <script src="../../../wct-browser-legacy/browser.js"></script>
+  <script type="module" src="../cookie-parser.js"></script>
 </head>
 <body>
-  <script>
+  <script type="module">
+  import '../../../url/url-polyfill.js';
+  import '../cookie-parser.js';
   /* global assert, Cookies */
   suite('Parse cookies', function() {
     let httpStr = 'rememberme=1; domain=foo.com;';
